+++ packages/manager/core/ovh-themes/ui-kit/ovh-components/icons/index.ts
@@ -1,3 +1,4 @@
-// TODO export * from './concept';
+export * from './concept';
 // TODO export * from './flags';
+export * from './ovh-font';
 export * from './system';