+++ packages/manager/modules/sms/src/sms/dashboard/telecom-sms-dashboard.module.js
@@ -2,6 +2,8 @@ import angular from 'angular';
 import '@uirouter/angularjs';
 import ovhManagerAdvices from '@ovh-ux/manager-advices';
 
+import creditTransfer from './creditTransfer';
+
 import component from './telecom-sms-dashboard.component';
 import routing from './routing';
 
@@ -10,7 +12,7 @@ import './telecom-sms-dashboard.scss';
 const moduleName = 'ovhManagerSmsDashboardModule';
 
 angular
-  .module(moduleName, ['ui.router', ovhManagerAdvices])
+  .module(moduleName, ['ui.router', creditTransfer, ovhManagerAdvices])
   .component(component.name, component)
   .config(routing)
   .run(/* @ngTranslationsInject:json ./translations */);