+++ packages/manager/modules/sms/src/sms/telecom-sms.constant.js
@@ -12,33 +12,14 @@ const SMS_GUIDES = {
       label: 'sms_guides_section_compose',
       guides: [
         {
-          label: 'sms_guides_compose_faq',
-          url: {
-            fr: 'https://www.ovh.com/fr/g2169.{title}',
-            es: 'https://docs.ovh.com/es/sms/',
-            en: 'https://docs.ovh.com/gb/en/sms/',
-            it: 'https://docs.ovh.com/it/sms/',
-            pl: 'https://docs.ovh.com/pl/sms/',
-          },
-        },
-        {
-          label: 'sms_guides_compose_filters',
-          url: {
-            fr: 'https://www.ovh.com/fr/g747.{title}',
-            es: 'https://docs.ovh.com/es/sms/',
-            en: 'https://docs.ovh.com/gb/en/sms/',
-            it: 'https://docs.ovh.com/it/sms/',
-            pl: 'https://docs.ovh.com/pl/sms/',
-          },
-        },
-        {
           label: 'sms_guides_compose_manager',
           url: {
-            fr: 'https://www.ovh.com/fr/g2142.{title}',
-            es:
-              'https://docs.ovh.com/es/sms/enviar-sms-desde-el-area-de-cliente/',
+            fr:
+              'https://docs.ovh.com/fr/sms/envoyer_des_sms_depuis_mon_espace_client/',
             en:
               'https://docs.ovh.com/gb/en/sms/send_sms_messages_via_control_panel/',
+            es:
+              'https://docs.ovh.com/es/sms/enviar-sms-desde-el-area-de-cliente/',
             it: 'https://docs.ovh.com/it/sms/inviare_sms_dallo_spazio_cliente/',
             pl:
               'https://docs.ovh.com/pl/sms/wysylanie-wiadomosci-sms-z-panelu-klienta/',
@@ -104,8 +85,12 @@ const SMS_GUIDES = {
         {
           label: 'sms_guides_outgoing_faq',
           url: {
-            fr: 'http://guides.ovh.com/SmsFaxFaqHistoriqueSms',
+            fr: 'https://docs.ovh.com/fr/sms/gerer-l-historique-des-sms/',
             es: 'https://docs.ovh.com/es/sms/historial-de-envios/',
+            en: 'https://docs.ovh.com/gb/en/sms/manage-sms-history/',
+            it: 'https://docs.ovh.com/it/sms/gestire-la-cronologia-degli-sms/',
+            pl:
+              'https://docs.ovh.com/pl/sms/zarzadzanie-historia-wiadomosci-sms/',
           },
         },
       ],
@@ -116,7 +101,16 @@ const SMS_GUIDES = {
         {
           label: 'sms_guides_options_callback',
           url: {
-            fr: 'http://guides.ovh.com/TelSmsCallBack',
+            fr:
+              'https://docs.ovh.com/fr/sms/tout_savoir_sur_les_utilisateurs_sms/#etape-5-specifier-une-url-de-callback',
+            es:
+              'https://docs.ovh.com/es/sms/usuarios-de-sms/#5-especificar-una-url-de-callback',
+            en:
+              'https://docs.ovh.com/gb/en/sms/everything_you_need_to_know_about_sms_users/#step-5-specify-a-callback-url',
+            it:
+              'https://docs.ovh.com/it/sms/tutto_sugli_utenti_sms/#step-5-specifica-un-url-di-callback',
+            pl:
+              'https://docs.ovh.com/pl/sms/informacje-o-uzytkownikach-sms/#etap-5-okreslanie-adresu-url-wywolania-zwrotnego',
           },
         },
       ],
@@ -127,27 +121,28 @@ const SMS_GUIDES = {
         {
           label: 'sms_guides_users_about',
           url: {
-            fr: 'https://www.ovh.com/fr/g2144.{title}',
-          },
-        },
-        {
-          label: 'sms_guides_users_faq',
-          url: {
-            fr: 'http://guides.ovh.com/SmsUserFaq',
+            fr:
+              'https://docs.ovh.com/fr/sms/tout_savoir_sur_les_utilisateurs_sms/',
+            es: 'https://docs.ovh.com/es/sms/usuarios-de-sms/',
+            en:
+              'https://docs.ovh.com/gb/en/sms/everything_you_need_to_know_about_sms_users/',
+            it: 'https://docs.ovh.com/it/sms/tutto_sugli_utenti_sms/',
+            pl: 'https://docs.ovh.com/pl/sms/informacje-o-uzytkownikach-sms/',
           },
         },
         {
           label: 'sms_guides_users_add',
           url: {
-            fr: 'http://guides.ovh.com/SmsUserAdd',
-          },
-        },
-        {
-          label: 'sms_guides_users_quota',
-          url: {
-            fr: 'http://guides.ovh.com/SmsUserQuota',
+            fr:
+              'https://docs.ovh.com/fr/sms/tout_savoir_sur_les_utilisateurs_sms/#etape-1-creer-un-utilisateur-api',
             es:
-              'https://docs.ovh.com/es/sms/enviar-sms-desde-correo-electronico/',
+              'https://docs.ovh.com/es/sms/usuarios-de-sms/#1-crear-un-usuario-de-la-api',
+            en:
+              'https://docs.ovh.com/gb/en/sms/everything_you_need_to_know_about_sms_users/#step-1-create-an-api-user',
+            it:
+              'https://docs.ovh.com/it/sms/tutto_sugli_utenti_sms/#step-1-crea-un-utente-api',
+            pl:
+              'https://docs.ovh.com/pl/sms/informacje-o-uzytkownikach-sms/#etap-1-utworzenie-uzytkownika-api',
           },
         },
       ],
@@ -212,6 +207,36 @@ const SMS_GUIDES = {
   ],
 };
 
+const SMPP_GUIDES = {
+  sections: [
+    {
+      label: 'sms_guides_section_smpp',
+      guides: [
+        {
+          label: 'sms_guides_smpp_control_panel',
+          url: {
+            fr: 'https://docs.ovh.com/fr/sms/smpp-control-panel/',
+            en: 'https://docs.ovh.com/gb/en/sms/smpp-control-panel/',
+            es: 'https://docs.ovh.com/es/sms/smpp-control-panel/',
+            it: 'https://docs.ovh.com/it/sms/smpp-control-panel/',
+            pl: 'https://docs.ovh.com/pl/sms/smpp-control-panel/',
+          },
+        },
+        {
+          label: 'sms_guides_smpp_specifications',
+          url: {
+            fr: 'https://docs.ovh.com/fr/sms/smpp-specifications/',
+            en: 'https://docs.ovh.com/gb/en/sms/smpp-specifications/',
+            es: 'https://docs.ovh.com/es/sms/smpp-specifications/',
+            it: 'https://docs.ovh.com/it/sms/smpp-specifications/',
+            pl: 'https://docs.ovh.com/pl/sms/smpp-specifications/',
+          },
+        },
+      ],
+    },
+  ],
+};
+
 const SMS_ALERTS = {
   sms: {
     bodyMaxLength: 70,
@@ -233,6 +258,7 @@ const SMS_PHONEBOOKS = {
 };
 
 export default {
+  SMPP_GUIDES,
   SMS_URL,
   SMS_GUIDES,
   SMS_ALERTS,