+++ packages/components/ovh-at-internet/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.3](https://github.com/ovh/manager/compare/@ovh-ux/ovh-at-internet@0.3.2...@ovh-ux/ovh-at-internet@0.3.3) (2023-02-07)
+
+**Note:** Version bump only for package @ovh-ux/ovh-at-internet
+
+
+
+
+
 ## [0.3.2](https://github.com/ovh/manager/compare/@ovh-ux/ovh-at-internet@0.3.1...@ovh-ux/ovh-at-internet@0.3.2) (2023-01-19)
 
 **Note:** Version bump only for package @ovh-ux/ovh-at-internet