+++ packages/manager/apps/dbaas-logs/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.9.8](https://github.com/ovh/manager/compare/@ovh-ux/manager-dbaas-logs-app@1.9.7...@ovh-ux/manager-dbaas-logs-app@1.9.8) (2023-02-08)
+
+**Note:** Version bump only for package @ovh-ux/manager-dbaas-logs-app
+
+
+
+
+
+## [1.9.7](https://github.com/ovh/manager/compare/@ovh-ux/manager-dbaas-logs-app@1.9.6...@ovh-ux/manager-dbaas-logs-app@1.9.7) (2023-02-07)
+
+**Note:** Version bump only for package @ovh-ux/manager-dbaas-logs-app
+
+
+
+
+
 ## [1.9.6](https://github.com/ovh/manager/compare/@ovh-ux/manager-dbaas-logs-app@1.9.5...@ovh-ux/manager-dbaas-logs-app@1.9.6) (2023-02-01)
 
 **Note:** Version bump only for package @ovh-ux/manager-dbaas-logs-app