+++ packages/manager/core/ovh-chakra-showcase/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.2.0](https://github.com/ovh/manager/compare/ovh-chakra-showcase@0.1.2...ovh-chakra-showcase@0.2.0) (2023-02-07)
+
+
+### Features
+
+* move server sidebar to container ([f446458](https://github.com/ovh/manager/commit/f446458542308f1ba0ae2cbccf9ed198081b5daa))
+
+
+
+
+
 ## [0.1.2](https://github.com/ovh/manager/compare/ovh-chakra-showcase@0.1.1...ovh-chakra-showcase@0.1.2) (2023-01-04)
 
 **Note:** Version bump only for package ovh-chakra-showcase