+++ packages/manager/modules/netapp/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.5.1](https://github.com/ovh/manager/compare/@ovh-ux/manager-netapp@1.5.0...@ovh-ux/manager-netapp@1.5.1) (2023-02-07)
+
+**Note:** Version bump only for package @ovh-ux/manager-netapp
+
+
+
+
+
 # [1.5.0](https://github.com/ovh/manager/compare/@ovh-ux/manager-netapp@1.4.3...@ovh-ux/manager-netapp@1.5.0) (2023-01-24)